+++ backend/src/main/java/hanglog/community/presentation/CommunityController.java
@@ -7,9 +7,9 @@ import hanglog.auth.domain.Accessor;
 import hanglog.community.dto.response.CommunityTripListResponse;
 import hanglog.community.dto.response.RecommendTripListResponse;
 import hanglog.community.service.CommunityService;
+import hanglog.trip.dto.response.TripDetailResponse;
 import hanglog.trip.dto.response.TripLedgerResponse;
 import hanglog.trip.service.LedgerService;
-import hanglog.trip.dto.response.TripDetailResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
