+++ backend/src/main/java/hanglog/community/service/CommunityService.java
@@ -7,18 +7,18 @@ import static hanglog.trip.domain.type.PublishedStatusType.PUBLISHED;
 import hanglog.auth.domain.Accessor;
 import hanglog.city.domain.City;
 import hanglog.city.domain.repository.CityRepository;
-import hanglog.community.domain.LikeInfo;
 import hanglog.community.domain.recommendstrategy.RecommendStrategies;
 import hanglog.community.domain.recommendstrategy.RecommendStrategy;
-import hanglog.community.domain.repository.LikeRepository;
 import hanglog.community.dto.CityElement;
 import hanglog.community.dto.CityElements;
-import hanglog.community.dto.LikeElement;
-import hanglog.community.dto.LikeElements;
 import hanglog.community.dto.response.CommunityTripListResponse;
 import hanglog.community.dto.response.CommunityTripResponse;
 import hanglog.community.dto.response.RecommendTripListResponse;
 import hanglog.global.exception.BadRequestException;
+import hanglog.like.domain.LikeInfo;
+import hanglog.like.dto.LikeElement;
+import hanglog.like.dto.LikeElements;
+import hanglog.like.repository.LikeRepository;
 import hanglog.trip.domain.Trip;
 import hanglog.trip.domain.repository.PublishedTripRepository;
 import hanglog.trip.domain.repository.TripCityRepository;
