+++ backend/src/test/java/hanglog/community/presentation/LikeControllerTest.java
@@ -15,9 +15,10 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import hanglog.auth.domain.MemberTokens;
-import hanglog.community.dto.request.LikeRequest;
-import hanglog.community.service.LikeService;
+import hanglog.like.dto.request.LikeRequest;
+import hanglog.like.service.LikeService;
 import hanglog.global.ControllerTest;
+import hanglog.like.presentation.LikeController;
 import jakarta.servlet.http.Cookie;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
