+++ backend/src/test/java/hanglog/integration/service/LikeServiceIntegrationTest.java
@@ -3,9 +3,9 @@ package hanglog.integration.service;
 import static hanglog.integration.IntegrationFixture.TRIP_CREATE_REQUEST;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
-import hanglog.community.domain.repository.LikeRepository;
-import hanglog.community.dto.request.LikeRequest;
-import hanglog.community.service.LikeService;
+import hanglog.like.repository.LikeRepository;
+import hanglog.like.dto.request.LikeRequest;
+import hanglog.like.service.LikeService;
 import hanglog.trip.infrastructure.CustomDayLogRepositoryImpl;
 import hanglog.trip.infrastructure.CustomTripCityRepositoryImpl;
 import hanglog.trip.service.TripService;
